refactor(plugin): hand cross-environment module loading to plugin-rsc
Two hand-rolled mechanisms did what plugin-rsc now does natively: dev-time
Proxy objects wrapping environment runners, and a virtual dist-importer
that rewrote import specifiers in renderChunk. Both are replaced by
import.meta.viteRsc.loadModule(), which plugin-rsc compiles to a runner
import in dev and to a relative import of the target environment's entry
chunk in build.

Both arguments are passed explicitly. loadModule() otherwise falls back to
the environment's first rollup input, and for ssr that is Universal
Deploy's `index` node entry rather than our ssr runtime ? it would have
loaded the wrong module silently. The arguments must also stay literals,
because plugin-rsc evaluates them statically from the source text.

This retires the hardcoded dist path fixed as a stopgap in the previous
commit: plugin-rsc derives the filename itself, and the emitted bundle now
contains import("../../rsc/index.js") with no __VIRTUAL_BUILD_ENTRY__
placeholder left anywhere in dist.

Drop the four manual @vitejs/plugin-rsc/vendor/react-server-dom/* entries
from optimizeDeps. plugin-rsc selects which react-server-dom build each
environment binds to; naming its vendored paths duplicated that decision
and pinned an internal layout.

Import the plugin from the package root rather than the /plugin deep path,
and add @vitejs/plugin-rsc/types to tsconfig so import.meta.viteRsc is
typed instead of implicitly any.

Remove vite6BuilderApp, which is unconditionally true in vike 0.4.260, and
the explicit useBuildAppHook: true, which has been the default since
Vite 7.

Fix the virtual module id spelled "enviroment-name" across all thirteen
importers. It is internal and not exported, so this is a rename with no
consumer impact.
