+++ packages/vike-react-rsc/src/runtime/virtual.d.ts
@@ -6,7 +6,7 @@ declare module "virtual:runtime/ssr" {
   const ssr: typeof import("./ssr");
   export = ssr;
 }
-declare module "virtual:enviroment-name" {
+declare module "virtual:environment-name" {
   const name: "rsc" | "ssr" | "client";
   export = name;
 }
